+++ src/RouteConfigSource/IRoutesConfigSource.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Butterfly\Adapter\Sf2Routing\RouteConfigSource;
+
+interface IRoutesConfigSource
+{
+    /**
+     * @return array
+     */
+    public function getConfig();
+}