+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+2.0.0
+---------
+
+Added functionality:
+
+ * Added annotations configuration support
+ * Added twig extension for url generate
+
 1.2.2
 -----